Rename SimpleFileKvStore to FileKvStore and refactor storage engine implementation
