--- examples/shakespeare_rag_comparison_improved.rs
@@ -1,817 +1,0 @@
-#![cfg(feature = "rag_examples")]
-//! Improved Shakespeare RAG vs GraphRAG Comparison Demo
-//!
-//! This example demonstrates enhanced RAG and GraphRAG implementations with:
-//! - Better semantic embeddings using TF-IDF and Shakespeare-specific features
-//! - Improved entity extraction for characters, themes, and relationships
-//! - More sophisticated knowledge graph construction
-//! - Comprehensive performance and quality analysis
-
-use oxidb::core::rag::document::Document;
-use oxidb::core::rag::embedder::{EmbeddingModel, SemanticEmbedder, TfIdfEmbedder};
-use oxidb::core::rag::graphrag::{GraphRAGEngineImpl, GraphRAGContext, KnowledgeNode, KnowledgeEdge};
-use oxidb::core::rag::retriever::{InMemoryRetriever, SimilarityMetric};
-use oxidb::core::rag::{GraphRAGEngine, Retriever};
-use oxidb::Value;
-use regex::Regex;
-use std::collections::HashMap;
-use std::fs;
-use std::path::Path;
-use std::time::{Duration, Instant};
-
-/// Shakespeare work metadata
-#[derive(Debug, Clone)]
-struct ShakespeareWork {
-    title: String,
-    url: String,
-    filename: String,
-    genre: String,
-}
-
-/// Enhanced performance metrics
-#[derive(Debug, Clone)]
-struct PerformanceMetrics {
-    retrieval_time: Duration,
-    processing_time: Duration,
-    result_count: usize,
-    relevance_score: f64,
-    semantic_quality: f64,
-    entity_count: usize,
-    relationship_count: usize,
-}
-
-/// Comprehensive comparison results
-#[derive(Debug)]
-struct ComparisonResult {
-    query: String,
-    rag_metrics: PerformanceMetrics,
-    graphrag_metrics: PerformanceMetrics,
-    rag_results: Vec<String>,
-    graphrag_results: Vec<String>,
-    quality_analysis: QualityAnalysis,
-}
-
-/// Quality analysis metrics
-#[derive(Debug)]
-struct QualityAnalysis {
-    context_relevance: f64,
-    character_coverage: f64,
-    theme_identification: f64,
-    relationship_accuracy: f64,
-}
-
-#[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    println!("🎭 Enhanced Shakespeare RAG vs GraphRAG Comparison");
-    println!("==================================================");
-
-    // Step 1: Setup and download Shakespeare works
-    let download_dir = "shakespeare_texts";
-    setup_download_directory(download_dir)?;
-
-    println!("\n📚 Downloading Shakespeare works...");
-    let works = get_shakespeare_works();
-    let mut raw_documents = Vec::new();
-    
-    for work in &works {
-        match download_shakespeare_work(work, download_dir).await {
-            Ok(content) => {
-                let processed_docs = process_shakespeare_text(&work.title, &content, &work.genre);
-                raw_documents.extend(processed_docs);
-                println!("✅ Downloaded and processed: {}", work.title);
-            }
-            Err(e) => {
-                println!("⚠️  Failed to download {}: {}", work.title, e);
-                let sample_docs = create_sample_shakespeare_content(&work.title, &work.genre);
-                raw_documents.extend(sample_docs);
-            }
-        }
-    }
-
-    println!("📄 Total documents processed: {}", raw_documents.len());
-
-    // Step 2: Create enhanced embeddings
-    println!("\n🧠 Creating enhanced embeddings...");
-    
-    // Create TF-IDF embedder for traditional RAG
-    let tfidf_embedder = TfIdfEmbedder::new(&raw_documents);
-    
-    // Create semantic embedder for GraphRAG
-    let semantic_embedder = SemanticEmbedder::new(512);
-    
-    // Generate embeddings for RAG system
-    println!("🔢 Generating TF-IDF embeddings for RAG...");
-    let mut rag_documents = Vec::new();
-    for doc in &raw_documents {
-        let embedding = tfidf_embedder.embed_document(doc).await?;
-        rag_documents.push(doc.clone().with_embedding(embedding));
-    }
-    
-    // Generate embeddings for GraphRAG system
-    println!("🎨 Generating semantic embeddings for GraphRAG...");
-    let mut graphrag_documents = Vec::new();
-    for doc in &raw_documents {
-        let embedding = semantic_embedder.embed_document(doc).await?;
-        graphrag_documents.push(doc.clone().with_embedding(embedding));
-    }
-
-    // Step 3: Setup enhanced RAG system
-    println!("\n🔍 Setting up enhanced RAG system...");
-    let rag_retriever = Box::new(InMemoryRetriever::new(rag_documents.clone()));
-
-    // Step 4: Setup enhanced GraphRAG system
-    println!("🕸️  Setting up enhanced GraphRAG system...");
-    let graphrag_retriever = Box::new(InMemoryRetriever::new(graphrag_documents.clone()));
-    let mut graphrag_engine = GraphRAGEngineImpl::new(graphrag_retriever);
-    
-    // Build enhanced knowledge graph
-    println!("🏗️  Building enhanced knowledge graph...");
-    graphrag_engine.build_knowledge_graph(&graphrag_documents).await?;
-    enhance_shakespeare_knowledge_graph(&mut graphrag_engine).await?;
-
-    // Step 5: Run comprehensive comparisons
-    println!("\n⚡ Running enhanced performance comparisons...");
-    let test_queries = get_enhanced_test_queries();
-    let mut results = Vec::new();
-
-    for query in &test_queries {
-        println!("\n🔎 Testing query: '{}'", query);
-        
-        // Test traditional RAG
-        let rag_result = benchmark_enhanced_rag_retrieval(&*rag_retriever, query, &semantic_embedder).await?;
-        
-        // Test GraphRAG
-        let graphrag_result = benchmark_enhanced_graphrag_retrieval(&graphrag_engine, query, &semantic_embedder).await?;
-        
-        // Analyze quality
-        let quality_analysis = analyze_result_quality(query, &rag_result.1, &graphrag_result.1);
-        
-        let comparison = ComparisonResult {
-            query: query.clone(),
-            rag_metrics: rag_result.0,
-            graphrag_metrics: graphrag_result.0,
-            rag_results: rag_result.1,
-            graphrag_results: graphrag_result.1,
-            quality_analysis,
-        };
-        
-        results.push(comparison);
-    }
-
-    // Step 6: Display enhanced analysis
-    println!("\n📊 Enhanced Analysis Results");
-    println!("============================");
-    
-    for result in &results {
-        display_enhanced_comparison_result(result);
-    }
-
-    // Step 7: Generate comprehensive summary
-    display_enhanced_summary(&results);
-
-    // Step 8: Demonstrate GraphRAG-specific capabilities
-    println!("\n🎯 GraphRAG-Specific Enhanced Capabilities");
-    println!("==========================================");
-    demonstrate_enhanced_graphrag_features(&graphrag_engine).await?;
-
-    println!("\n✅ Enhanced Shakespeare RAG vs GraphRAG comparison completed!");
-    println!("📈 Results show improved semantic understanding and relationship detection!");
-
-    Ok(())
-}
-
-/// Setup download directory
-fn setup_download_directory(dir: &str) -> Result<(), Box<dyn std::error::Error>> {
-    if !Path::new(dir).exists() {
-        fs::create_dir_all(dir)?;
-    }
-    Ok(())
-}
-
-/// Get Shakespeare works to download
-fn get_shakespeare_works() -> Vec<ShakespeareWork> {
-    vec![
-        ShakespeareWork {
-            title: "Romeo and Juliet".to_string(),
-            url: "https://www.gutenberg.org/files/1513/1513-0.txt".to_string(),
-            filename: "romeo_juliet.txt".to_string(),
-            genre: "Tragedy".to_string(),
-        },
-        ShakespeareWork {
-            title: "Hamlet".to_string(),
-            url: "https://www.gutenberg.org/files/1524/1524-0.txt".to_string(),
-            filename: "hamlet.txt".to_string(),
-            genre: "Tragedy".to_string(),
-        },
-        ShakespeareWork {
-            title: "Macbeth".to_string(),
-            url: "https://www.gutenberg.org/files/1533/1533-0.txt".to_string(),
-            filename: "macbeth.txt".to_string(),
-            genre: "Tragedy".to_string(),
-        },
-        ShakespeareWork {
-            title: "A Midsummer Night's Dream".to_string(),
-            url: "https://www.gutenberg.org/files/1514/1514-0.txt".to_string(),
-            filename: "midsummer.txt".to_string(),
-            genre: "Comedy".to_string(),
-        },
-    ]
-}
-
-/// Download Shakespeare work
-async fn download_shakespeare_work(
-    work: &ShakespeareWork,
-    download_dir: &str,
-) -> Result<String, Box<dyn std::error::Error>> {
-    let file_path = Path::new(download_dir).join(&work.filename);
-    
-    if file_path.exists() {
-        return Ok(fs::read_to_string(file_path)?);
-    }
-
-    // Use reqwest for HTTP downloads
-    let client = reqwest::Client::builder()
-        .timeout(Duration::from_secs(30))
-        .build()?;
-    
-    let response = client.get(&work.url).send().await?;
-    let content = response.text().await?;
-    
-    fs::write(&file_path, &content)?;
-    Ok(content)
-}
-
-/// Process Shakespeare text into documents
-fn process_shakespeare_text(title: &str, content: &str, genre: &str) -> Vec<Document> {
-    let mut documents = Vec::new();
-    
-    // Clean the text (remove Project Gutenberg headers/footers)
-    let cleaned_content = clean_gutenberg_text(content);
-    
-    // Split into acts and scenes
-    let act_scene_regex = Regex::new(r"(?i)(ACT\s+[IVX]+|SCENE\s+[IVX]+)").unwrap();
-    let parts: Vec<&str> = act_scene_regex.split(&cleaned_content).collect();
-    
-    let mut current_act = "Unknown";
-    let mut current_scene = "Unknown";
-    
-    for (_i, part) in parts.iter().enumerate() {
-        if part.trim().is_empty() {
-            continue;
-        }
-        
-        // Determine if this is an act or scene marker
-        if let Some(captures) = act_scene_regex.find(part) {
-            let marker = captures.as_str().to_uppercase();
-            if marker.starts_with("ACT") {
-                current_act = "ACT";
-                current_scene = "Scene 1"; // Reset scene
-            } else if marker.starts_with("SCENE") {
-                current_scene = "SCENE";
-            }
-            continue;
-        }
-        
-        // Create document for this part
-        if part.len() > 100 { // Only include substantial content
-            let doc_id = format!("{}_{}_{}_{}", 
-                title.to_lowercase().replace(' ', "_"),
-                genre.to_lowercase(),
-                current_act.to_lowercase().replace(' ', "_"),
-                current_scene.to_lowercase().replace(' ', "_")
-            );
-            
-            let mut metadata = HashMap::new();
-            metadata.insert("title".to_string(), Value::Text(title.to_string()));
-            metadata.insert("genre".to_string(), Value::Text(genre.to_string()));
-            metadata.insert("act".to_string(), Value::Text(current_act.to_string()));
-            metadata.insert("scene".to_string(), Value::Text(current_scene.to_string()));
-            
-            let document = Document::new(doc_id, part.trim().to_string())
-                .with_metadata(metadata);
-            
-            documents.push(document);
-        }
-    }
-    
-    // If no acts/scenes found, create chunks
-    if documents.is_empty() {
-        let chunks = chunk_text(&cleaned_content, 1000);
-        for (i, chunk) in chunks.iter().enumerate() {
-            let doc_id = format!("{}_{}_{}", 
-                title.to_lowercase().replace(' ', "_"),
-                genre.to_lowercase(),
-                i
-            );
-            
-            let mut metadata = HashMap::new();
-            metadata.insert("title".to_string(), Value::Text(title.to_string()));
-            metadata.insert("genre".to_string(), Value::Text(genre.to_string()));
-            metadata.insert("chunk_index".to_string(), Value::Integer(i as i64));
-            
-            let document = Document::new(doc_id, chunk.clone())
-                .with_metadata(metadata);
-            
-            documents.push(document);
-        }
-    }
-    
-    documents
-}
-
-/// Clean Project Gutenberg text
-fn clean_gutenberg_text(content: &str) -> String {
-    let lines: Vec<&str> = content.lines().collect();
-    let mut start_idx = 0;
-    let mut end_idx = lines.len();
-    
-    // Find start of actual content (after Project Gutenberg header)
-    for (i, line) in lines.iter().enumerate() {
-        if line.contains("*** START OF") || line.contains("ACT I") || line.contains("SCENE I") {
-            start_idx = i;
-            break;
-        }
-    }
-    
-    // Find end of actual content (before Project Gutenberg footer)
-    for (i, line) in lines.iter().enumerate().rev() {
-        if line.contains("*** END OF") || line.contains("THE END") {
-            end_idx = i;
-            break;
-        }
-    }
-    
-    lines[start_idx..end_idx].join("\n")
-}
-
-/// Chunk text into smaller pieces
-fn chunk_text(text: &str, max_chunk_size: usize) -> Vec<String> {
-    let words: Vec<&str> = text.split_whitespace().collect();
-    let mut chunks = Vec::new();
-    let mut current_chunk = Vec::new();
-    let mut current_size = 0;
-    
-    for word in words {
-        if current_size + word.len() > max_chunk_size && !current_chunk.is_empty() {
-            chunks.push(current_chunk.join(" "));
-            current_chunk.clear();
-            current_size = 0;
-        }
-        
-        current_chunk.push(word);
-        current_size += word.len() + 1; // +1 for space
-    }
-    
-    if !current_chunk.is_empty() {
-        chunks.push(current_chunk.join(" "));
-    }
-    
-    chunks
-}
-
-/// Create sample content for testing
-fn create_sample_shakespeare_content(title: &str, genre: &str) -> Vec<Document> {
-    let sample_contents = vec![
-        "Romeo and Juliet meet at the Capulet party and fall in love at first sight.",
-        "Hamlet sees his father's ghost who tells him of his murder by Claudius.",
-        "Macbeth meets the three witches who prophesy he will become king.",
-        "The lovers in the forest are confused by Puck's magical interventions.",
-    ];
-    
-    sample_contents.iter().enumerate().map(|(i, content)| {
-        let doc_id = format!("{}_sample_{}", title.to_lowercase().replace(' ', "_"), i);
-        let mut metadata = HashMap::new();
-        metadata.insert("title".to_string(), Value::Text(title.to_string()));
-        metadata.insert("genre".to_string(), Value::Text(genre.to_string()));
-        metadata.insert("is_sample".to_string(), Value::Text("true".to_string()));
-        
-        Document::new(doc_id, content.to_string()).with_metadata(metadata)
-    }).collect()
-}
-
-/// Get enhanced test queries
-fn get_enhanced_test_queries() -> Vec<String> {
-    vec![
-        "love and romance in Shakespeare".to_string(),
-        "tragic deaths and endings".to_string(),
-        "family conflicts and feuds".to_string(),
-        "supernatural elements and ghosts".to_string(),
-        "power and ambition themes".to_string(),
-        "comedy and humor".to_string(),
-        "betrayal and revenge".to_string(),
-        "character relationships and bonds".to_string(),
-        "fate versus free will".to_string(),
-        "honor and nobility".to_string(),
-    ]
-}
-
-/// Benchmark enhanced RAG retrieval
-async fn benchmark_enhanced_rag_retrieval(
-    retriever: &InMemoryRetriever,
-    query: &str,
-    embedder: &SemanticEmbedder,
-) -> Result<(PerformanceMetrics, Vec<String>), Box<dyn std::error::Error>> {
-    let start_time = Instant::now();
-    
-    // Create query document and embed it
-    let query_doc = Document::new("query".to_string(), query.to_string());
-    let query_embedding = embedder.embed_document(&query_doc).await?;
-    
-    let retrieval_start = Instant::now();
-    let documents = retriever.retrieve(&query_embedding, 10, SimilarityMetric::Cosine).await?;
-    let retrieval_time = retrieval_start.elapsed();
-    
-    let processing_time = start_time.elapsed();
-    
-    // Calculate enhanced metrics
-    let relevance_score = calculate_relevance_score(query, &documents);
-    let semantic_quality = calculate_semantic_quality(query, &documents);
-    
-    let results: Vec<String> = documents.iter()
-        .map(|doc| format!("{}: {}", doc.id, doc.content.chars().take(100).collect::<String>()))
-        .collect();
-    
-    let metrics = PerformanceMetrics {
-        retrieval_time,
-        processing_time,
-        result_count: documents.len(),
-        relevance_score,
-        semantic_quality,
-        entity_count: 0, // RAG doesn't extract entities
-        relationship_count: 0,
-    };
-    
-    Ok((metrics, results))
-}
-
-/// Benchmark enhanced GraphRAG retrieval
-async fn benchmark_enhanced_graphrag_retrieval(
-    engine: &GraphRAGEngineImpl,
-    query: &str,
-    embedder: &SemanticEmbedder,
-) -> Result<(PerformanceMetrics, Vec<String>), Box<dyn std::error::Error>> {
-    let start_time = Instant::now();
-    
-    // Create query document and embed it
-    let query_doc = Document::new("query".to_string(), query.to_string());
-    let query_embedding = embedder.embed_document(&query_doc).await?;
-    
-    let retrieval_start = Instant::now();
-    let context = GraphRAGContext {
-        query_embedding,
-        max_hops: 2,
-        min_confidence: 0.3,
-        include_relationships: vec![],
-        exclude_relationships: vec![],
-        entity_types: vec!["CHARACTER".to_string(), "THEME".to_string()],
-    };
-    
-    let result = engine.retrieve_with_graph(context).await?;
-    let retrieval_time = retrieval_start.elapsed();
-    
-    let processing_time = start_time.elapsed();
-    
-    // Calculate enhanced metrics
-    let relevance_score = calculate_relevance_score(query, &result.documents);
-    let semantic_quality = calculate_semantic_quality(query, &result.documents);
-    
-    let results: Vec<String> = result.documents.iter()
-        .map(|doc| format!("{}: {}", doc.id, doc.content.chars().take(100).collect::<String>()))
-        .collect();
-    
-    let metrics = PerformanceMetrics {
-        retrieval_time,
-        processing_time,
-        result_count: result.documents.len(),
-        relevance_score,
-        semantic_quality,
-        entity_count: result.relevant_entities.len(),
-        relationship_count: result.entity_relationships.len(),
-    };
-    
-    Ok((metrics, results))
-}
-
-/// Calculate relevance score
-fn calculate_relevance_score(query: &str, documents: &[Document]) -> f64 {
-    if documents.is_empty() {
-        return 0.0;
-    }
-    
-    let query_lower = query.to_lowercase();
-    let query_words: Vec<&str> = query_lower.split_whitespace().collect();
-    let mut total_relevance = 0.0;
-    
-    for doc in documents {
-        let doc_lower = doc.content.to_lowercase();
-        let doc_words: Vec<&str> = doc_lower.split_whitespace().collect();
-        let mut matches = 0;
-        
-        for query_word in &query_words {
-            if doc_words.iter().any(|&word| word.contains(query_word) || query_word.contains(word)) {
-                matches += 1;
-            }
-        }
-        
-        let doc_relevance = matches as f64 / query_words.len() as f64;
-        total_relevance += doc_relevance;
-    }
-    
-    total_relevance / documents.len() as f64
-}
-
-/// Calculate semantic quality
-fn calculate_semantic_quality(_query: &str, documents: &[Document]) -> f64 {
-    if documents.is_empty() {
-        return 0.0;
-    }
-    
-    // Simple semantic quality based on content diversity and relevance
-    let mut unique_themes = std::collections::HashSet::new();
-    let mut total_quality = 0.0;
-    
-    for doc in documents {
-        // Check for thematic diversity
-        let content_lower = doc.content.to_lowercase();
-        if content_lower.contains("love") || content_lower.contains("romance") {
-            unique_themes.insert("love");
-        }
-        if content_lower.contains("death") || content_lower.contains("die") {
-            unique_themes.insert("death");
-        }
-        if content_lower.contains("power") || content_lower.contains("king") {
-            unique_themes.insert("power");
-        }
-        if content_lower.contains("family") || content_lower.contains("father") {
-            unique_themes.insert("family");
-        }
-        
-        // Quality based on content length and structure
-        let content_quality = (doc.content.len() as f64 / 1000.0).min(1.0);
-        total_quality += content_quality;
-    }
-    
-    let theme_diversity = unique_themes.len() as f64 / 4.0; // Max 4 themes
-    let avg_content_quality = total_quality / documents.len() as f64;
-    
-    (theme_diversity + avg_content_quality) / 2.0
-}
-
-/// Analyze result quality
-fn analyze_result_quality(query: &str, rag_results: &[String], graphrag_results: &[String]) -> QualityAnalysis {
-    let context_relevance = calculate_context_relevance(query, rag_results, graphrag_results);
-    let character_coverage = calculate_character_coverage(rag_results, graphrag_results);
-    let theme_identification = calculate_theme_identification(query, rag_results, graphrag_results);
-    let relationship_accuracy = calculate_relationship_accuracy(rag_results, graphrag_results);
-    
-    QualityAnalysis {
-        context_relevance,
-        character_coverage,
-        theme_identification,
-        relationship_accuracy,
-    }
-}
-
-/// Calculate context relevance
-fn calculate_context_relevance(query: &str, rag_results: &[String], graphrag_results: &[String]) -> f64 {
-    let query_lower = query.to_lowercase();
-    let rag_relevance = rag_results.iter()
-        .map(|r| if r.to_lowercase().contains(&query_lower) { 1.0 } else { 0.0 })
-        .sum::<f64>() / rag_results.len().max(1) as f64;
-    
-    let graphrag_relevance = graphrag_results.iter()
-        .map(|r| if r.to_lowercase().contains(&query_lower) { 1.0 } else { 0.0 })
-        .sum::<f64>() / graphrag_results.len().max(1) as f64;
-    
-    (rag_relevance + graphrag_relevance) / 2.0
-}
-
-/// Calculate character coverage
-fn calculate_character_coverage(rag_results: &[String], graphrag_results: &[String]) -> f64 {
-    let characters = vec!["romeo", "juliet", "hamlet", "macbeth", "othello"];
-    let all_results = [rag_results, graphrag_results].concat();
-    
-    let covered_characters = characters.iter()
-        .filter(|&&character| {
-            all_results.iter().any(|result| result.to_lowercase().contains(character))
-        })
-        .count();
-    
-    covered_characters as f64 / characters.len() as f64
-}
-
-/// Calculate theme identification
-fn calculate_theme_identification(query: &str, rag_results: &[String], graphrag_results: &[String]) -> f64 {
-    let themes = vec!["love", "death", "power", "revenge", "family"];
-    let query_lower = query.to_lowercase();
-    let all_results = [rag_results, graphrag_results].concat();
-    
-    let relevant_themes = themes.iter()
-        .filter(|&&theme| query_lower.contains(theme))
-        .count();
-    
-    if relevant_themes == 0 {
-        return 0.5; // Neutral score if no specific themes in query
-    }
-    
-    let identified_themes = themes.iter()
-        .filter(|&&theme| {
-            query_lower.contains(theme) && 
-            all_results.iter().any(|result| result.to_lowercase().contains(theme))
-        })
-        .count();
-    
-    identified_themes as f64 / relevant_themes as f64
-}
-
-/// Calculate relationship accuracy
-fn calculate_relationship_accuracy(_rag_results: &[String], graphrag_results: &[String]) -> f64 {
-    // GraphRAG should identify more relationships
-    let relationship_indicators = vec!["loves", "kills", "betrays", "serves", "fights"];
-    
-    let relationship_count = relationship_indicators.iter()
-        .map(|&indicator| {
-            graphrag_results.iter()
-                .filter(|result| result.to_lowercase().contains(indicator))
-                .count()
-        })
-        .sum::<usize>();
-    
-    (relationship_count as f64 / 10.0).min(1.0) // Normalize to 0-1
-}
-
-/// Display enhanced comparison result
-fn display_enhanced_comparison_result(result: &ComparisonResult) {
-    println!("\n📝 Query: '{}'", result.query);
-    println!("   RAG Performance:");
-    println!("     ⏱️  Retrieval time: {:?}", result.rag_metrics.retrieval_time);
-    println!("     🔄 Processing time: {:?}", result.rag_metrics.processing_time);
-    println!("     📊 Results count: {}", result.rag_metrics.result_count);
-    println!("     🎯 Relevance score: {:.3}", result.rag_metrics.relevance_score);
-    println!("     🧠 Semantic quality: {:.3}", result.rag_metrics.semantic_quality);
-    
-    println!("   GraphRAG Performance:");
-    println!("     ⏱️  Retrieval time: {:?}", result.graphrag_metrics.retrieval_time);
-    println!("     🔄 Processing time: {:?}", result.graphrag_metrics.processing_time);
-    println!("     📊 Results count: {}", result.graphrag_metrics.result_count);
-    println!("     🎯 Relevance score: {:.3}", result.graphrag_metrics.relevance_score);
-    println!("     🧠 Semantic quality: {:.3}", result.graphrag_metrics.semantic_quality);
-    println!("     👥 Entities found: {}", result.graphrag_metrics.entity_count);
-    println!("     🔗 Relationships: {}", result.graphrag_metrics.relationship_count);
-    
-    println!("   📈 Quality Analysis:");
-    println!("     🎯 Context relevance: {:.3}", result.quality_analysis.context_relevance);
-    println!("     👥 Character coverage: {:.3}", result.quality_analysis.character_coverage);
-    println!("     🎭 Theme identification: {:.3}", result.quality_analysis.theme_identification);
-    println!("     🔗 Relationship accuracy: {:.3}", result.quality_analysis.relationship_accuracy);
-    
-    if !result.rag_results.is_empty() {
-        println!("   📄 RAG Sample: {}", result.rag_results[0].chars().take(80).collect::<String>());
-    }
-    if !result.graphrag_results.is_empty() {
-        println!("   🕸️  GraphRAG Sample: {}", result.graphrag_results[0].chars().take(80).collect::<String>());
-    }
-}
-
-/// Display enhanced summary
-fn display_enhanced_summary(results: &[ComparisonResult]) {
-    println!("\n📈 Enhanced Performance Summary");
-    println!("==============================");
-    
-    let avg_rag_time: f64 = results.iter()
-        .map(|r| r.rag_metrics.retrieval_time.as_nanos() as f64)
-        .sum::<f64>() / results.len() as f64;
-    
-    let avg_graphrag_time: f64 = results.iter()
-        .map(|r| r.graphrag_metrics.retrieval_time.as_nanos() as f64)
-        .sum::<f64>() / results.len() as f64;
-    
-    let avg_rag_relevance: f64 = results.iter()
-        .map(|r| r.rag_metrics.relevance_score)
-        .sum::<f64>() / results.len() as f64;
-    
-    let avg_graphrag_relevance: f64 = results.iter()
-        .map(|r| r.graphrag_metrics.relevance_score)
-        .sum::<f64>() / results.len() as f64;
-    
-    let avg_rag_quality: f64 = results.iter()
-        .map(|r| r.rag_metrics.semantic_quality)
-        .sum::<f64>() / results.len() as f64;
-    
-    let avg_graphrag_quality: f64 = results.iter()
-        .map(|r| r.graphrag_metrics.semantic_quality)
-        .sum::<f64>() / results.len() as f64;
-    
-    println!("⏱️  Average Retrieval Times:");
-    println!("   RAG: {:.2}ms", avg_rag_time / 1_000_000.0);
-    println!("   GraphRAG: {:.2}ms", avg_graphrag_time / 1_000_000.0);
-    
-    println!("🎯 Average Relevance Scores:");
-    println!("   RAG: {:.3}", avg_rag_relevance);
-    println!("   GraphRAG: {:.3}", avg_graphrag_relevance);
-    
-    println!("🧠 Average Semantic Quality:");
-    println!("   RAG: {:.3}", avg_rag_quality);
-    println!("   GraphRAG: {:.3}", avg_graphrag_quality);
-    
-    let speed_factor = if avg_rag_time < avg_graphrag_time {
-        avg_graphrag_time / avg_rag_time
-    } else {
-        avg_rag_time / avg_graphrag_time
-    };
-    
-    let quality_improvement = ((avg_graphrag_relevance - avg_rag_relevance) / avg_rag_relevance) * 100.0;
-    
-    println!("\n🏆 Key Insights:");
-    if avg_rag_time < avg_graphrag_time {
-        println!("   📈 RAG is {:.1}x faster than GraphRAG", speed_factor);
-    } else {
-        println!("   📈 GraphRAG is {:.1}x faster than RAG", speed_factor);
-    }
-    
-    if quality_improvement > 0.0 {
-        println!("   🎯 GraphRAG shows {:.1}% better relevance", quality_improvement);
-    } else {
-        println!("   🎯 RAG shows {:.1}% better relevance", -quality_improvement);
-    }
-}
-
-/// Enhance Shakespeare knowledge graph
-async fn enhance_shakespeare_knowledge_graph(
-    engine: &mut GraphRAGEngineImpl,
-) -> Result<(), Box<dyn std::error::Error>> {
-    // Add well-known character relationships
-    let character_relationships = vec![
-        ("ROMEO", "JULIET", "LOVES", 0.95),
-        ("HAMLET", "CLAUDIUS", "SEEKS_REVENGE", 0.90),
-        ("MACBETH", "LADY_MACBETH", "MARRIED_TO", 0.92),
-        ("OTHELLO", "DESDEMONA", "LOVES", 0.88),
-        ("OTHELLO", "IAGO", "BETRAYED_BY", 0.85),
-    ];
-    
-    for (char1, char2, relationship, confidence) in character_relationships {
-        // Create entities if they don't exist
-        let entity1 = KnowledgeNode {
-            id: 0, // Will be assigned
-            entity_type: "CHARACTER".to_string(),
-            name: char1.to_string(),
-            description: Some(format!("Shakespeare character: {}", char1)),
-            embedding: None,
-            properties: HashMap::new(),
-            confidence_score: 0.95,
-        };
-        
-        let entity2 = KnowledgeNode {
-            id: 0, // Will be assigned
-            entity_type: "CHARACTER".to_string(),
-            name: char2.to_string(),
-            description: Some(format!("Shakespeare character: {}", char2)),
-            embedding: None,
-            properties: HashMap::new(),
-            confidence_score: 0.95,
-        };
-        
-        let id1 = engine.add_entity(entity1).await?;
-        let id2 = engine.add_entity(entity2).await?;
-        
-        let edge = KnowledgeEdge {
-            id: 0, // Will be assigned
-            from_entity: id1,
-            to_entity: id2,
-            relationship_type: relationship.to_string(),
-            description: Some(format!("{} {} {}", char1, relationship.to_lowercase(), char2)),
-            confidence_score: confidence,
-            weight: Some(confidence),
-        };
-        
-        engine.add_relationship(edge).await?;
-    }
-    
-    Ok(())
-}
-
-/// Demonstrate enhanced GraphRAG features
-async fn demonstrate_enhanced_graphrag_features(
-    _engine: &GraphRAGEngineImpl,
-) -> Result<(), Box<dyn std::error::Error>> {
-    println!("🔍 Character relationship analysis:");
-    
-    // This is a simplified demonstration since we can't easily query specific entities
-    // In a full implementation, you would have methods to find entities by name
-    println!("   • Enhanced entity extraction identifies characters, themes, and locations");
-    println!("   • Relationship detection finds love, conflict, and family connections");
-    println!("   • Confidence scoring provides quality metrics for relationships");
-    println!("   • Multi-hop reasoning enables complex query answering");
-    
-    println!("\n🎭 Theme and character associations:");
-    println!("   • ROMEO ↔ LOVE (confidence: 0.95)");
-    println!("   • HAMLET ↔ REVENGE (confidence: 0.90)");
-    println!("   • MACBETH ↔ AMBITION (confidence: 0.88)");
-    
-    println!("\n🏰 Location and setting connections:");
-    println!("   • VERONA ← Romeo and Juliet");
-    println!("   • ELSINORE ← Hamlet");
-    println!("   • SCOTLAND ← Macbeth");
-    
-    Ok(())
-}
