+++ src/core/storage/engine/implementations/file/mod.rs
@@ -1,0 +1,7 @@
+//! Module for the `FileKvStore` implementation.
+
+mod persistence; // internal
+mod recovery;
+pub mod store;
+
+pub use store::FileKvStore;
