+++ src/core/storage/engine/implementations/mod.rs
@@ -1,8 +1,8 @@
 pub mod in_memory; // Keep in_memory declaration
-pub mod simple_file; // Add simple_file module
+pub mod file; // Renamed from simple_file
 
 pub use in_memory::InMemoryKvStore;
-pub use simple_file::SimpleFileKvStore; // Expose SimpleFileKvStore
+pub use file::FileKvStore; // Expose FileKvStore
 
 #[cfg(test)]
 mod tests;
