--- src/core/storage/engine/implementations/simple_file/mod.rs
@@ -1,7 +1,0 @@
-//! Module for the `SimpleFileKvStore` implementation.
-
-mod persistence; // Will be pub(super)
-mod recovery;
-pub mod store; // Will be pub(super)
-
-pub use store::SimpleFileKvStore;
