+++ src/core/storage/engine/implementations/tests/file_tests.rs
@@ -1,0 +1,2 @@
+// Reuse the existing test implementations from the migrated file
+include!("./simple_file_tests.rs");
