+++ src/core/storage/engine/implementations/tests/mod.rs
@@ -1,3 +1,3 @@
 // This file will declare test modules for different storage engine implementations.
 pub mod in_memory_tests;
-pub mod simple_file_tests;
+pub mod file_tests;
