+++ src/core/storage/engine/mod.rs
@@ -9,5 +9,5 @@ pub mod wal;
 pub use buffer_pool_manager::BufferPoolManager;
 pub use disk_manager::DiskManager;
 pub use implementations::in_memory::InMemoryKvStore;
-pub use implementations::simple_file::SimpleFileKvStore;
+pub use implementations::file::FileKvStore;
 pub use page::{Page, PageHeader, PageType, PAGE_SIZE};
